requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
chromadb==0.4.18
pydantic==2.4.2
groq==0.4.0
pytest==7.4.3
jinja2==3.1.2
numpy==1.26.0
tqdm==4.66.1
validators==0.22.0